refactor: remove old promotion codes implemented by WidestType_t
