+++ infini_train/src/kernels/cpu/fill.cc
@@ -5,7 +5,7 @@
 #include "infini_train/include/dtype_dispatch.h"
 #include "infini_train/include/tensor.h"
 
-#include "infini_train/src/core/cpu/cpu_dispatch.h"
+#include "infini_train/src/core/runtime/cpu/cpu_dispatch.h"
 
 namespace infini_train::kernels::cpu {
 void Fill(std::shared_ptr<Tensor> tensor, double value) {
